+++ src/components/Navbar.jsx
@@ -5,7 +5,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 const Navbar = () => (
   <ul className="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion" id="accordionSidebar">
 
-    <a className="sidebar-brand d-flex align-items-center justify-content-center" href="index.html">
+    <a className="sidebar-brand d-flex align-items-center justify-content-center" href={`${process.env.PUBLIC_URL}/`}>
       <div className="sidebar-brand-icon rotate-n-15">
         <i className="fas fa-laugh-wink" />
       </div>