+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [Unreleased]
 
+  - Introduce React Router, refs #4
   - Setup CI, refs #2
   - Display pool token price, refs #7